+++ ElectronNET.Host/api/browserWindows.ts
@@ -706,6 +706,27 @@ export = (socket: SocketIO.Socket, app: Electron.App) => {
         getWindowById(id).setVibrancy(type);
     });
 
+    socket.on('browserWindowAddExtension', (path) => {
+        const extensionName = BrowserWindow.addExtension(path);
+
+        electronSocket.emit('browserWindow-addExtension-completed', extensionName);
+    });
+
+    socket.on('browserWindowRemoveExtension', (name) => {
+        BrowserWindow.removeExtension(name);
+    });
+
+    socket.on('browserWindowGetExtensions', () => {
+        const extensionsList = BrowserWindow.getExtensions();
+        const chromeExtensionInfo = [];
+
+        Object.keys(extensionsList).forEach(key => {
+            chromeExtensionInfo.push(extensionsList[key]);
+        });
+
+        electronSocket.emit('browserWindow-getExtensions-completed', chromeExtensionInfo);
+    });
+
     function getWindowById(id: number): Electron.BrowserWindow {
         for (let index = 0; index < windows.length; index++) {
             const element = windows[index];